requirements.txt
requests>=2.28.0
dnspython>=2.3.0
python-whois>=0.8.0
colorama>=0.4.6
beautifulsoup4>=4.11.0
shodan>=1.28.0
ssl-cert-check>=0.1.0
